106/ Refactor Tax Lot Endpoint - moved db access from service to repo
