+++ src/tax-lot/tax-lot.service.ts
@@ -5,23 +5,7 @@ import { DB, DbType } from "src/global/providers/db.provider";
 import { TaxLotRepository } from "./tax-lot.repository";
 import { FeatureFlagConfig } from "src/config";
 import { ConfigType } from "@nestjs/config";
-import {
-  taxLot,
-  zoningDistrict,
-  zoningDistrictClass,
-  zoningDistrictZoningDistrictClass,
-} from "src/schema";
-import { eq, sql } from "drizzle-orm";
-import {
-  DataRetrievalException,
-  ResourceNotFoundException,
-} from "src/exception";
-import {
-  SelectTaxLot,
-  SelectTaxLotNested,
-  SelectTaxLotSpatial,
-} from "src/schema/tax-lot";
-import { MultiPolygon } from "geojson";
+import { TaxLotRepo } from "./tax-lot.repo";
 
 @Injectable()
 export class TaxLotService {
@@ -32,42 +16,16 @@ export class TaxLotService {
     @Inject(DB)
     private readonly db: DbType,
 
+    @Inject(TaxLotRepo)
+    private readonly taxLotRepo: TaxLotRepo,
+
     @Inject(FeatureFlagConfig.KEY)
     private featureFlagConfig: ConfigType<typeof FeatureFlagConfig>,
   ) {}
 
-  #checkTaxLotByBbl = this.db.query.taxLot
-    .findFirst({
-      columns: {
-        bbl: true,
-      },
-      where: (taxLot, { eq, sql }) => eq(taxLot.bbl, sql.placeholder("bbl")),
-    })
-    .prepare("checkTaxLotByBbl");
-
   async findTaxLotByBbl(bbl: string) {
     if (this.featureFlagConfig.useDrizzle) {
-      let result: SelectTaxLotNested | undefined;
-      try {
-        result = await this.db.query.taxLot.findFirst({
-          columns: {
-            boroughId: false,
-            landUseId: false,
-            wgs84: false,
-            liFt: false,
-          },
-          where: (taxLot, { eq }) => eq(taxLot.bbl, bbl),
-          with: {
-            borough: true,
-            landUse: true,
-          },
-        });
-      } catch {
-        throw new DataRetrievalException();
-      }
-
-      if (result === undefined) throw new ResourceNotFoundException();
-      return result;
+      return await this.taxLotRepo.taxLotByBbl(bbl);
     } else {
       return this.taxLotRepository.findOne(
         { bbl },
@@ -81,44 +39,7 @@ export class TaxLotService {
 
   async findTaxLotByBblGeoJson(bbl: string) {
     if (this.featureFlagConfig.useDrizzle) {
-      let result: SelectTaxLotSpatial | undefined;
-      try {
-        result = await this.db.query.taxLot.findFirst({
-          columns: {
-            bbl: true,
-            block: true,
-            lot: true,
-            address: true,
-          },
-          extras: {
-            geometry: sql<string>`ST_AsGeoJSON(${taxLot.wgs84}, 6)`.as(
-              "geometry",
-            ),
-          },
-          where: (taxLot, { eq }) => eq(taxLot.bbl, bbl),
-          with: {
-            borough: true,
-            landUse: true,
-          },
-        });
-      } catch {
-        throw new DataRetrievalException();
-      }
-      if (result === undefined) throw new ResourceNotFoundException();
-      const geometry = JSON.parse(result.geometry) as MultiPolygon;
-      return {
-        type: "Feature",
-        id: result.bbl,
-        properties: {
-          bbl: result.bbl,
-          borough: result.borough,
-          block: result.block,
-          lot: result.lot,
-          address: result.address,
-          landUse: result.landUse,
-        },
-        geometry,
-      };
+      return await this.taxLotRepo.taxLotByBblGeoJson(bbl);
     } else {
       return this.taxLotRepository.findOne({ bbl });
     }
@@ -126,32 +47,7 @@ export class TaxLotService {
 
   async findZoningDistrictByTaxLotBbl(bbl: string) {
     if (this.featureFlagConfig.useDrizzle) {
-      let taxLotCheck: Pick<SelectTaxLot, "bbl"> | undefined;
-      try {
-        taxLotCheck = await this.#checkTaxLotByBbl.execute({ bbl });
-      } catch {
-        throw new DataRetrievalException();
-      }
-      if (taxLotCheck === undefined) throw new ResourceNotFoundException();
-      try {
-        const zoningDistricts = await this.db
-          .select({
-            id: zoningDistrict.id,
-            label: zoningDistrict.label,
-          })
-          .from(zoningDistrict)
-          .leftJoin(
-            taxLot,
-            sql`ST_Intersects(${taxLot.liFt}, ${zoningDistrict.liFt})`,
-          )
-          .where(eq(taxLot.bbl, bbl));
-
-        return {
-          zoningDistricts,
-        };
-      } catch {
-        throw new DataRetrievalException();
-      }
+      return await this.taxLotRepo.zoningDistrictByTaxLotBbl(bbl);
     } else {
       throw new Error(
         "Zoning district by tax lot bbl route not implemented in Mikro orm",
@@ -161,49 +57,7 @@ export class TaxLotService {
 
   async findZoningDistrictClassByTaxLotBbl(bbl: string) {
     if (this.featureFlagConfig.useDrizzle) {
-      let taxLotCheck: Pick<SelectTaxLot, "bbl"> | undefined;
-      try {
-        taxLotCheck = await this.#checkTaxLotByBbl.execute({ bbl });
-      } catch {
-        throw new DataRetrievalException();
-      }
-      if (taxLotCheck === undefined) throw new ResourceNotFoundException();
-      try {
-        const zoningDistrictClasses = await this.db
-          .select({
-            id: zoningDistrictClass.id,
-            category: zoningDistrictClass.category,
-            description: zoningDistrictClass.description,
-            url: zoningDistrictClass.url,
-            color: zoningDistrictClass.color,
-          })
-          .from(zoningDistrictClass)
-          .leftJoin(
-            zoningDistrictZoningDistrictClass,
-            eq(
-              zoningDistrictZoningDistrictClass.zoningDistrictClassId,
-              zoningDistrictClass.id,
-            ),
-          )
-          .leftJoin(
-            zoningDistrict,
-            eq(
-              zoningDistrictZoningDistrictClass.zoningDistrictId,
-              zoningDistrict.id,
-            ),
-          )
-          .leftJoin(
-            taxLot,
-            sql`ST_Intersects(${taxLot.liFt}, ${zoningDistrict.liFt})`,
-          )
-          .where(eq(taxLot.bbl, bbl));
-
-        return {
-          zoningDistrictClasses,
-        };
-      } catch {
-        throw new DataRetrievalException();
-      }
+      return await this.taxLotRepo.zoningDistrictClassByTaxLotBbl(bbl);
     } else {
       throw new Error(
         "Zoning district class by tax lot bbl route not implemented in Mikro orm",
